Sequence Generator with singleton
Singleton implementation which will generate a sequence of numbers to the callers a. The idea here is that there is only a single number/sequence generator and all the numbers follow a perfect sequence.
b. So when we do a call to the generator like getNextNumber() we should get the next number in the sequence no matter how we obtained the generator.