implement a common file with singleton
Create a Singleton implementation which will write to a common file.
a. Imagine that you wanted to audit all the actions in your application.
b. It would make sense to log that into a file with a guarantee that all
callers will write to the same file.(For each call to the FileAuditManager your code will generate a
timestamp and will write the message entry into the file with a
newline character at the end of each message.)